Map / sampling point data layer: added marker label
